builder improvements
Expands the test coverage and uses the real aws builder class in the tests. In a
way this is trading one evil for another, since we're now using java reflection
to access private variables on the builder, but it captures and tests the intent
of the code better than it did before.